feat: 旅行を編集ボタンの表示方法を変更
「旅行を編集」ボタンの表示方法を画面埋め込みからモーダルに変更しました。
